pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dont-track-me"
version = "0.1.0"
description = "Modular anti-tracking toolkit — audit and protect your digital privacy"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dns = ["dnspython>=2.6"]
metadata = ["Pillow>=10.0", "pypdf>=4.0"]
social-api = ["keyring>=25.0"]
reddit = ["dont-track-me[social-api]"]
youtube = ["dont-track-me[social-api]"]
all = ["dont-track-me[dns,metadata,social-api]"]
dev = [
    "dont-track-me[all]",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[project.scripts]
dtm = "dont_track_me.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/dont_track_me"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort (import ordering)
    "N",     # pep8-naming
    "UP",    # pyupgrade (modern Python idioms)
    "B",     # flake8-bugbear (common pitfalls)
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",  # line too long — handled by formatter
    "N818",  # exception name suffix — AuthenticationRequired is clear enough
]

[tool.ruff.lint.isort]
known-first-party = ["dont_track_me"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
